api-integration/omie/requirements.txt
functions-framework==3.*
datetime~=4.4
google-cloud-logging
google-cloud-bigquery
facebook_business
dataclass-wizard
pandas==1.4.*


# requests~=2.27.1